move blog post fetching into data folder since its a data source for project and not a util
